refactor(enterprise): localize legacy share mapping
