+++ packages/enterprise/package.json
@@ -27,6 +27,7 @@
     "@hono/standard-validator": "catalog:",
     "hono": "catalog:",
     "hono-openapi": "catalog:",
+    "effect": "catalog:",
     "js-base64": "3.7.7",
     "luxon": "catalog:",
     "nitro": "3.0.1-alpha.1",
